\docType{package}
\name{tbdbaid-package}
\alias{tbdbaid}
\alias{tbdbaid-package}
\title{tbdbaid: a collection of functions for querying TBdb.}
\description{
This package provides a collection of functions for making
standardized queries against TBdb, TBdbPlus, and
occassionally the Denver Health Data Warehouse.
}
\author{
Matt Parker \email{mattmparker@gmail.com}
}
\keyword{package}
\keyword{tbdb}

